requirements.txt
asgiref==3.2.10
attrs==20.2.0
autobahn==20.12.3
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
Django==3.1.4
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.12.2
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
mysqlclient==2.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2020.1
regex==2020.10.15
requests==2.25.1
screen==1.0.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
txaio==20.4.1
tzlocal==2.1
urllib3==1.25.8
webencodings==0.5.1
xlwt==1.3.0
zope.interface==5.1.2
